Data.Array.Byte

Description

Compatibility layer for Data.Array.Byte, providing boxed wrappers for ByteArray# and MutableByteArray# and relevant instances for GHC < 9.4. Include it into your Cabal file:

build-depends: base
if impl(ghc < 9.4)
  build-depends: data-array-byte

and then import Data.Array.Byte unconditionally.

Originally derived from primitive package.

Documentation

data ByteArray

Boxed wrapper for ByteArray#.

Since ByteArray# is an unlifted type and not a member of kind Type, things like [ByteArray#] or IO ByteArray# are ill-typed. To work around this inconvenience this module provides a standard boxed wrapper, inhabiting Type. Clients are expected to use ByteArray in higher-level APIs, but wrap and unwrap ByteArray internally as they please and use functions from GHC.Exts.

Constructors

ByteArray ByteArray#

data MutableByteArray s

Boxed wrapper for MutableByteArray#.

Since MutableByteArray# is an unlifted type and not a member of kind Type, things like [MutableByteArray#] or IO MutableByteArray# are ill-typed. To work around this inconvenience this module provides a standard boxed wrapper, inhabiting Type. Clients are expected to use MutableByteArray in higher-level APIs, but wrap and unwrap MutableByteArray internally as they please and use functions from GHC.Exts.

Constructors

MutableByteArray (MutableByteArray# s)